Added option to disable mines remaining detection.
This option is added since the new interface of
https://freeminesweeper.org does not have correct display of the number
of mines remaining on the upper left corner of the board any more. I
decide to convey the total number of mines to the program before the
beginning of a game so that the program can keep track of the number of
mines remaining itself.